Prevent preview of next token ending in 10 (revised)
Refactored code from previous commit to use BigNumber.isZero(). Also removed some extraneous code.

Original commit notes below:

When the token ID ends in 10, this one and the next will not go to the minter. Currently the preview image is misleading for these tokens because the user expects the next token that is previewed. This code removes the preview for tokens ending in 10 (the Lil Nounder's reward) and instead displays a question mark image.